pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-contentguard"
version = "0.1.0"
description = "Content rights adjudication environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
server = "contentguard.server.app:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["contentguard", "contentguard.server"]
package-dir = { "contentguard" = ".", "contentguard.server" = "server" }